fixed some meta issues
